+++ src/waveform_viewer.ts
@@ -12,14 +12,19 @@ import Semaphore = require('ts-semaphore');
 const b64decode = (str: string): string => Buffer.from(str, 'base64').toString('binary');
 const b64encode = (str: string): string => Buffer.from(str, 'binary').toString('base64');
 
+const  cmdAccess = new Semaphore(1);
 
 abstract class BaseViewer {
+
     protected viewerProcess: null | ChildProcess = null;
     protected spawnArgs: Array<string>;
     protected executable: string
     protected validWavesExtension: Array<string>;
     protected context: ExtensionContext;
 
+    protected _followWavesEnabled : boolean = false;
+    protected setupDone : boolean = false; // Is set to true to avoid calling setup multiple time.
+    public abstract readonly canFollowWaveSelection : boolean;
     
     public events = new EventEmitter(); //! Events used to interface with the viewer.
 
@@ -29,9 +34,9 @@ abstract class BaseViewer {
     public verboseLog: boolean = false;
 
     private periodicCalls: NodeJS.Timeout[] = []
-    protected stdoutChunks: Array<string> = [];
+    protected llGetRemaining: string = "";
 
-    private cmdAccess = new Semaphore(1);
+    //private cmdAccess = new Semaphore(1);
 
     protected stdoutCb: (args: WaveformViewerCbArgs) => Promise<void>;
 
@@ -40,10 +45,19 @@ abstract class BaseViewer {
         this.spawnArgs = execArgs;
         this.validWavesExtension = validWavesFiles;
         this.context = context;
+        
         this.stdoutCb = stdoutCallBack;
-        console.log(`Viewer construction done, chunks are ${this.stdoutChunks !== undefined}`);
-
         commands.executeCommand("setContext", "diplomat-host:viewerEnabled", false);
+    }
+
+    protected setUpCapabilities() : void
+    {
+        if(! this.setupDone){
+            this.setupDone = true;
+            this._followWavesEnabled = this.canFollowWaveSelection;
+            console.log(`Set "follow" capability to ${this.canFollowWaveSelection}`);
+            commands.executeCommand("setContext", "diplomat-host:viewerCanFollowSignal", this.canFollowWaveSelection);
+        }
     }
 
     abstract openWave(wavefile: string): void;
@@ -53,15 +67,12 @@ abstract class BaseViewer {
 
     protected forwardToCallback(data: any) {
 
-        let command = this.stdoutChunks.join("");
-        this.stdoutChunks = [];
-
         let cbArgs: WaveformViewerCbArgs | undefined = undefined;
         try {
             cbArgs = assertEquals<WaveformViewerCbArgs>(data);
         } catch (error) {
             if (this.verboseLog) {
-                console.log("Failed to parse viewer output\n%s", command);
+                console.log("Failed to parse viewer output\n%s", data);
                 cbArgs = undefined;
             }
         }
@@ -73,7 +84,7 @@ abstract class BaseViewer {
     }
 
     protected async ensureClosed() {
-
+        
         this.stopAllPeriodic();
         commands.executeCommand("diplomat-host.waves.clear-annotations");
         
@@ -91,6 +102,8 @@ abstract class BaseViewer {
         }
 
         await this.viewerClosed;
+        // Disable all context
+        this.llGetRemaining = "";
         commands.executeCommand("setContext", "diplomat-host:viewerEnabled", false);
     }
 
@@ -109,6 +122,7 @@ abstract class BaseViewer {
 
     private _lowLevelSendData(cmd :string) {
         if (this.viewerProcess !== null) {
+
             if (!this.viewerProcess.stdin?.write(cmd + "\n")) {
                 return new Promise((resolve) => {
                     this.viewerProcess?.stdin?.once("drain", resolve);
@@ -119,39 +133,46 @@ abstract class BaseViewer {
 
     private async _lowLevelGetData() : Promise<string> {
         let buffer : string[] = [];
-        if (this.viewerProcess && this.viewerProcess.stdout && this.viewerProcess.stdout.readable)
+        if (this.viewerProcess?.stdout?.readable)
         {
             
             let rawData : null | Buffer = this.viewerProcess.stdout.read();
-            while(buffer.length == 0 || ! buffer[buffer.length-1].endsWith("§"))
+            let endFound : boolean = false;
+            while(!endFound)
             {
                 
                 let data : string = "";
-                let trimmed : string = "";
+
                 if(rawData !== null)
-                    {
+                {
                         data = rawData.toString("utf8");
-                        trimmed = data.trimEnd();
-                        // console.log(`LL Got ${data}`);
-                        // console.log(`LL Got Stuff`);
-                        buffer.push(trimmed.endsWith('§') ? trimmed : data);
+
+                        let position : integer;
+                        if((position = data.indexOf("§")) == -1)
+                        {
+                            buffer.push(data);
+                        }
+                        else {
+                            buffer.push(data.slice(0,position));
+                            endFound = true;
+                        }
                 }
 
-                if(rawData === null || ! trimmed.endsWith('§'))
+                if(rawData === null || ! endFound)
                     rawData = (await once(this.viewerProcess.stdout,"data"))[0];
                 // else
                 //     console.log("Clean exit of LL get");
 
             } 
         }
-
-        let ret : string = buffer.join("").slice(0,-1);
+        let ret : string = buffer.join("").trimEnd();
+        await this._lowLevelFlushStdout();
         return ret;
     }
 
     private async _lowLevelFlushStdout() 
     {
-        if (this.viewerProcess && this.viewerProcess.stdout && this.viewerProcess.stdout.readable)
+        if (this.viewerProcess?.stdout?.readable)
         {
             let data;
             while((data = this.viewerProcess.stdout.read()) !== null)
@@ -162,15 +183,17 @@ abstract class BaseViewer {
 
 
     protected async sendCommand(cmd: string) {
-        await this.cmdAccess.use(() => {this._lowLevelSendData(cmd);});
+        await cmdAccess.use(async () => {await this._lowLevelSendData(cmd);});
     }
 
     protected async exchangeCommand(cmd: string) {
         
-        return await this.cmdAccess.use(
-            () => {return this._lowLevelFlushStdout();}
-        ).then(() => {return this._lowLevelSendData(cmd);}
-        ).then(() => {return this._lowLevelGetData()});
+        return await cmdAccess.use(
+           async () => {
+            await this._lowLevelFlushStdout();
+            await  this._lowLevelSendData(cmd);
+            return await this._lowLevelGetData();
+        });
         
     }
 
@@ -191,17 +214,37 @@ abstract class BaseViewer {
         (async () => {await this.ensureClosed();});        
     }
 
+    public setWaveFollow(newValue : boolean) {
+        if(this.canFollowWaveSelection)
+            this._followWavesEnabled = newValue;
+    }
+
+    public get waveFollowingEnabled() : boolean
+    {
+        return this._followWavesEnabled;
+    }
+
 };
 
 
 export class GTKWaveViewer extends BaseViewer {
 
+    public readonly canFollowWaveSelection : boolean = true;
 
     /**
      * openWave
      */
 
+
+    // protected setUpCapabilities() {
+    //     this._followWavesEnabled = this.canFollowWaveSelection;
+    //     console.log(`Set "follow" capability to ${this.canFollowWaveSelection}`);
+    //     commands.executeCommand("setContext", "diplomat-host:viewerCanFollowSignal", this.canFollowWaveSelection);
+    // }
+
     public async openWave(wavefile: string) {
+
+        this.setUpCapabilities();
         await this.ensureClosed();
         let arglist = workspace.getConfiguration("diplomatServer.tools.GTKWave").get<string[]>("options");
         
@@ -266,6 +309,20 @@ export class GTKWaveViewer extends BaseViewer {
                         commands.executeCommand("diplomat-host.waves.refresh-annotations");
                     }
                 });
+                
+        }, 500);
+
+        this.addPeriodicCommand(() => {
+            if(this._followWavesEnabled)
+            {
+                this.exchangeCommand("tell_selected")
+                    .then((result: string) => {
+                        if(result.trim() != "")
+                        {
+                            this.forwardToCallback(JSON.parse(result));
+                        }
+                    });
+            }
                 
         }, 500);